backend/requirements.txt
Django==5.2
djangorestframework==3.15.2
djoser==2.3.1
psycopg2-binary==2.9.10
gunicorn==20.1.0
python-dotenv==1.0.1

drf_extra_fields==3.7.0
django-filter==25.1
django_import_export==4.3.7
pillow==11.1.0
cryptography==44.0.1
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.3.1
dj-database-url==2.1.0
whitenoise==6.6.0
drf-spectacular==0.27.2
tablib==3.8.0
numpy==2.2.5

flake8==7.2.0
isort==6.0.1